chatroom deletion + last question (#68)
* added last question feature

* updated applicantchat.js and chat.js

- chatrooms deletes anon-user from db
- applicantchat now listens to whether anon-user exists in db and updates accordingly
- chat now removes listeners once user and chat are deleted
