+++ loan-calculator-client/package.json
@@ -13,10 +13,8 @@
     "react": "^16.13.1",
     "react-dom": "^16.13.1",
     "react-multiselect-checkboxes": "^0.1.1",
-    "react-new-window": "^0.1.2",
     "react-router-dom": "^5.2.0",
     "react-scripts": "3.4.3",
-    "react-splitter-layout": "^4.0.0",
     "react-tooltip": "^4.2.10"
   },
   "scripts": {
