+++ loan-calculator-client/src/App.js
@@ -15,7 +15,8 @@ import AdminSignup from './components/AdminSignup.js'
 import AdminLogin from './components/AdminLogin.js'
 import PrivateRoute from './components/PrivateRoute.js'
 import NavigationBar from './components/NavigationBar.js'
-import MediaPortal from './components/MediaPortal.js'
+import AdminChat from './components/AdminChat.js'
+import Video from './components/Video.js'
 import { ContextAPI } from './components/Context.js'
 import firebase from './scripts/firebase.js'
 
@@ -33,7 +34,7 @@ class App extends React.Component {
     this.handleAdminJoinRoom = this.handleAdminJoinRoom.bind(this)
   }
   componentDidMount() {
-    this.handleAdminCheck()  
+    this.handleAdminCheck()
   }
   render() {
     const value = {
@@ -56,7 +57,8 @@ class App extends React.Component {
               <Route path='/login' component={AdminLogin} />
               <Route path='/signup' component={AdminSignup} />
               <Route path='/calculator' component={Calculator} />
-              <PrivateRoute path='/chat' component={MediaPortal} />
+              <Route path='/chat' component={AdminChat} />
+              <Route path='/video' component={Video} />
               <PrivateRoute path='/chat-rooms' component={ChatRooms} />
               <Route path='/' component={Home} />
             </Switch>
