+++ loan-calculator-client/src/components/ApplicantChat.js
@@ -1,11 +1,8 @@
 import React from 'react';
 import Chat from './Chat.js';
-import MediaPortal from './MediaPortal.js'
-import NavigationBar from './NavigationBar.js'
 import firebase from '../scripts/firebase.js';
 import { Button, Modal } from 'antd';
-import { CloseOutlined, CaretUpOutlined, CaretDownOutlined, VideoCameraOutlined } from '@ant-design/icons';
-import NewWindow from 'react-new-window'
+import { CloseOutlined, CaretUpOutlined, CaretDownOutlined } from '@ant-design/icons';
 import "antd/dist/antd.css";
 import "../css/App.css";
 import "../css/ApplicantChat.css";
@@ -19,75 +16,37 @@ class ApplicantChat extends React.Component {
     * minimized-chat: there has been a chat window opened, but it is currently minimized
     * closed-chat: there is no chat window open, and the user can open one by clicking "Chat with an Advisor"
     */
-   super(props);
+    super(props);
     this.state = {
       chatOpen: false,
       chatMinimized: false,
       name: '',
-      isSignedIn: false,
-      videoPortal: false
-    }
-    this.activateVideoPortal = this.activateVideoPortal.bind(this)
+      isSignedIn: false
+    };
+    this.anonUserListener = null;
   }
 
-  render() {
-    const objectFeatures = {
-      width: "1800",
-      height: "1500"
-    }
-    if (!this.state.chatOpen) {
-      return (
-        <Button
-          className='open-chat-button'
-          onClick={() => {
-            this.setState({ chatOpen: true })
-            if (firebase.auth().currentUser !== null) {
-              this.setState({
-                isSignedIn: true,
-                chatMinimized: false
-              })
-            }
-          }}
-        >
-          Chat with an Advisor
-        </Button>
-      );
-    } else if (!this.state.isSignedIn) {
-      return (
-        <Modal
-          title="Enter your name below"
-          visible={this.state.chatOpen && !this.state.isSignedIn}
-          onOk={this.handleOk}
-          onCancel={() => this.setState({ chatOpen: false })}
-        >
-          <input placeholder="Name" name="name" value={this.state.name} onChange={(event) => this.setState({ name: event.target.value })}></input>
-        </Modal>
-      );
-    }
-    return (
-      <div className='user-chat-popup' id={this.state.chatMinimized ? 'minimized-chat' : 'open-chat'}>
-        <div className='user-chat-header'>
-          <Button icon={< VideoCameraOutlined />} type='text' size='small' shape='circle' onClick={this.activateVideoPortal}/>
-          {this.state.chatMinimized ? (
-            <Button icon={<CaretUpOutlined />} type='text' size='small' shape='circle' onClick={() => this.setState({ chatMinimized: false })} />
-          ) : (
-              <Button icon={<CaretDownOutlined />} type='text' size='small' shape='circle' onClick={() => this.setState({ chatMinimized: true })} />
-            )
+  async componentDidMount() {
+    if (firebase.auth().currentUser) {
+      let user = await firebase.auth().currentUser
+      const db = firebase.firestore()
+      let anonUserRef = db.collection('anon-users').doc(user.uid)
+      let anonUserSnapshot = await anonUserRef.get()
+      if (anonUserSnapshot.exists) {
+        this.anonUserListener = anonUserRef.onSnapshot(snapshot => {
+          if (!snapshot.data()) {
+            this.anonUserListener()
+            firebase.auth().signOut()
+            this.setState({
+              isSignedIn: false,
+              chatOpen: false
+            })
           }
-          <Button icon={<CloseOutlined />} type='text' size='small' shape='circle' onClick={() => this.setState({ chatOpen: false })} />
-        </div>
-        <div style={{ height: 'calc(100% - 20px)' }}>
-          <Chat/>
-        </div>
-        {this.state.videoPortal && (
-          <NewWindow features={objectFeatures} title="videoPortal" onUnload={() => this.setState({videoPortal: false})}>
-            <NavigationBar />
-            <MediaPortal/>
-          </NewWindow>
-        )}
-
-      </div>
-    )
+        })
+      } else {
+        firebase.auth().signOut()
+      }
+    }
   }
 
   // Saves anonymous user information to Firestore (name, uid, chatbot status)
@@ -125,7 +84,18 @@ class ApplicantChat extends React.Component {
 
   //This function will create another room in the database and send a greeting from the chatbot
   async createRoom(room_name, creator_name, creator_uid) {    
-    const db = firebase.firestore();    
+    const db = firebase.firestore();
+    let anonUserRef = await db.collection('anon-users').doc(creator_uid)
+    this.anonUserListener = anonUserRef.onSnapshot(snapshot => {
+      if (!snapshot.data()) {
+        this.anonUserListener()
+        this.setState({
+          isSignedIn: false,
+          chatOpen: false
+        })
+      }
+    })
+
     let roomRef = await db.collection('chat-rooms').doc(room_name)
     roomRef.set({
       "creator_name": creator_name,
@@ -145,11 +115,54 @@ class ApplicantChat extends React.Component {
     })
   }
 
-  activateVideoPortal() {
-    this.setState({videoPortal: true}, function() {
-      console.log('videoPortal: ', this.state.videoPortal)
-    })
-  }
+  render() {
+    if (!this.state.chatOpen) {
+      return (
+        <Button
+          className='open-chat-button'
+          onClick={() => {
+            this.setState({ chatOpen: true });
+            if (firebase.auth().currentUser !== null) {
+              this.setState({
+                isSignedIn: true,
+                chatMinimized: false
+              });
+            }
+          }}
+        >
+          Chat with an Advisor
+        </Button>
+      );
+    } else if (!this.state.isSignedIn) {
+      return (
+        <Modal
+          title="Enter your name below"
+          visible={this.state.chatOpen && !this.state.isSignedIn}
+          onOk={this.handleOk}
+          onCancel={() => this.setState({ chatOpen: false})}
+        >
+          <input placeholder="Name" name="name" value={this.state.name} onChange={(event) => this.setState({ name: event.target.value })}></input>
+        </Modal>
+      );
+    }
+    return (
+      <div className='user-chat-popup' id={this.state.chatMinimized ? 'minimized-chat' : 'open-chat'}>
+        <div className='user-chat-header'>
+          {this.state.chatMinimized ? (
+            <Button icon={<CaretUpOutlined />} type='text' size='small' shape='circle' onClick={() => this.setState({ chatMinimized: false })} />
+          ) : (
+              <Button icon={<CaretDownOutlined />} type='text' size='small' shape='circle' onClick={() => this.setState({ chatMinimized: true })} />
+            )
+          }
+          <Button icon={<CloseOutlined />} type='text' size='small' shape='circle' onClick={() => this.setState({ chatOpen: false })} />
+        </div>
+        <div style={{ height: 'calc(100% - 20px)' }}>
+          <Chat/>
+        </div>
+      </div>
+  
+    );
+  } 
 }
 
 export default ApplicantChat;
