+++ loan-calculator-client/src/css/Video.css
@@ -1,206 +1,13 @@
-/*
-	remoteVideo: Your partner's webcam view
-*/
-#remoteVideo {
-	/*border: solid red;*/
-	border-radius: 10px;
-	position: absolute;
-	top: 0px;
-	left: 0px;
-	width: 100%;
-	height: 100%;
-	background-color: black;
+div#videos {
+    display: flex;
+    flex-direction: row;
+    flex-wrap: wrap;
 }
 
-
-/*
-	localVideo: Your webcam view
-*/
-
-#localVideo {
-	border: solid white;
-	border-radius: 10px;
-	position: absolute;
-	bottom: 0px;
-	width: 20%;
-	right: 0px;
-}
-
-
-#hangupBtn {
-	border-color: red;
-	border-width: 2px; 
-	position: absolute;
-	left: 41.5%;
-	bottom: 0px;
-	width: 20%;
-	border-radius: 10px;
-	color: red;
-	cursor: pointer;
-	transition-duration: 0.4s;
-}
-
-#hangupBtn:hover {
-	background-color: #ffcccb;
-}
-
-#menu {
-	background-color: white;
-}
-
-#controlsToggleDivOpen {
-	position: relative;
-	display: inline-block;
-	height: 100%;
-	width: 40%;
-}
-
-#controlsToggleDivClosed {
-	position: relative;
-	display: inline-block;
-	height: 100%;
-	width: 100%;
-}
-#controlToggleBtn {
-	width: 100%;
-	height: 100%;
-}
-.controllersDivOpen {
-	position: absolute;
-	display: inline-block;
-	width: 60%;
-	height: 100%;
-}
-.controllersDivClosed {
-	visibility: hidden;
-	position: absolute;
-	display: inline-block;
-	width: 60%;
-	height: 100%;
-}
-.contollerBtn {
-	width: 100%;
-	height: 25%;
-	padding-top: 1px;
-	/*border: solid black;*/
-}
-
-/*Activate Camera & Mic Button CSS*/
-#cameraBtn {
-	width: 100%;
-	height: 100%;
-}
-#cameraTwoTone {
-	position: absolute;
-	top: 50%;
-	right: 46%;
-	font-size: 20px;
-}
-
-/*Join Room Btn CSS*/
-#joinRoomBtn {
-	width: 100%;
-	height: 100%;
-	text-align: center;	
-}
-#teamOutlined {
-	position: absolute;
-	top: 50%;
-	right: 46%;
-	font-size: 20px;	
-}
-
-/*Create Room Btn CSS*/
-#createRoomBtn {
-	width: 100%;
-	height: 100%;
-	text-align: center;
-}
-#plusCircleTwoTone {
-	position: absolute;
-	top: 50%;
-	right: 46%;
-	font-size: 20px;
-}
-
-/*Toggle Video Btn CSS*/
-#toggleVideoBtn {
-	width: 100%;
-	height: 100%;
-	text-align: center;
-}
-#videoCameraTwoTone {
-	position: absolute;
-	top: 50%;
-	right: 46%;
-	font-size: 20px;
-}
-
-/*Toggle Sound Btn CSS*/
-#toggleSoundBtn {
-	width: 100%;
-	height: 100%;
-	text-align: center;
-}
-#soundTwoTone {
-	position: absolute;
-	top: 50%;
-	right: 46%;
-	font-size: 20px;	
-}
-
-
-/*Toggle Chat Btn CSS*/
-#toggleChatBtn {
-	margin-left: 25%;
-	width: 50%;
-	text-align: center;	
-}
-
-
-/* Navigation Stuff */
-.media-portal {
-	/*border: solid black;*/
-	position: absolute;
-	width: 100%;
-	height: 94.8%;
-	top: 50px;
-
-}
-
-.sider {
-	background-color: cyan;
-}
-
-.video {
-	/*border: solid orange;*/
-	background-color: #f0f0f0;
-}
-
-.video-interface-card {
-	border-color: transparent;
-	box-shadow: 5px 15px 30px #383838;
-	background-color: #383838;
-	border-radius: 10px;
-	height: 90%;
-	width: 75%;
-	margin: auto;
-	position: relative;
-	margin-top: 20px;
-}
-
-.footer {
-	background-color: #528400;
-}
-
-.chat-interface {
-	background-color: pink;
-	width: 500px;
-	/*border-style: groove;*/
-}
-
-.controller-interface {
-	background-color: transparent;
-	width: 500px;
-	/*border-style: groove;*/
+div#videos > video {
+    background: black;
+    width: 640px;
+    height: 100%;
+    display: block;
+    margin: 1em;
 }
