+++ loan-calculator-client/src/scripts/chatbot.js
@@ -6,6 +6,7 @@ const redirectQuestion = "Chat with a financial advisor";
 const principalBalanceQuestion = "Why are the Principal Balance and the Payoff Amount Different?";
 const autoLoanQuestion = "How does this auto loan work?";
 const latePaymentQuestion = "How is this loan affected by late payments?";
+
 const interestResponse = "An interst payment is the payment from a borrow to a lender of an amount above repayment of the amount borrowed, at a particular rate.";
 const loanTermResponse = "The loan term is the length of time that you have to repay the loan."
 const principalBalanceResponse = "The Payoff Amount represents the principal balance, plus accrued interest and any applicable fees. This is the amount you would have to pay to fully pay off your car."
